Trait rusoto_mediatailor::MediaTailor

pub trait MediaTailor {
    fn delete_playback_configuration(
        &self, 
        input: DeletePlaybackConfigurationRequest
    ) -> RusotoFuture<DeletePlaybackConfigurationResponse, DeletePlaybackConfigurationError>;

    fn get_playback_configuration(
        &self, 
        input: GetPlaybackConfigurationRequest
    ) -> RusotoFuture<GetPlaybackConfigurationResponse, GetPlaybackConfigurationError>;

    fn list_playback_configurations(
        &self, 
        input: ListPlaybackConfigurationsRequest
    ) -> RusotoFuture<ListPlaybackConfigurationsResponse, ListPlaybackConfigurationsError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn put_playback_configuration(
        &self, 
        input: PutPlaybackConfigurationRequest
    ) -> RusotoFuture<PutPlaybackConfigurationResponse, PutPlaybackConfigurationError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<(), TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<(), UntagResourceError>;
}

Trait representing the capabilities of the MediaTailor API. MediaTailor clients implement this trait.

Required methods

fn delete_playback_configuration(
&self,
input: DeletePlaybackConfigurationRequest
) -> RusotoFuture<DeletePlaybackConfigurationResponse, DeletePlaybackConfigurationError>

Deletes the playback configuration for the specified name.

fn get_playback_configuration(
&self,
input: GetPlaybackConfigurationRequest
) -> RusotoFuture<GetPlaybackConfigurationResponse, GetPlaybackConfigurationError>

Returns the playback configuration for the specified name.

fn list_playback_configurations(
&self,
input: ListPlaybackConfigurationsRequest
) -> RusotoFuture<ListPlaybackConfigurationsResponse, ListPlaybackConfigurationsError>

Returns a list of the playback configurations defined in AWS Elemental MediaTailor. You can specify a maximum number of configurations to return at a time. The default maximum is 50. Results are returned in pagefuls. If MediaTailor has more configurations than the specified maximum, it provides parameters in the response that you can use to retrieve the next pageful.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

Returns a list of the tags assigned to the specified playback configuration resource.

fn put_playback_configuration(
&self,
input: PutPlaybackConfigurationRequest
) -> RusotoFuture<PutPlaybackConfigurationResponse, PutPlaybackConfigurationError>

Adds a new playback configuration to AWS Elemental MediaTailor.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<(), TagResourceError>

Adds tags to the specified playback configuration resource. You can specify one or more tags to add.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<(), UntagResourceError>

Removes tags from the specified playback configuration resource. You can specify one or more tags to remove.

Loading content...